recommendor/requirements.txt
Click==7.0
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
SQLAlchemy==1.3.3
Werkzeug==0.15.4


lxml==4.6.2
bs4==0.0.1
requests==2.25.0
requests-oauthlib==1.3.0
urllib3==1.26.2
pandas==1.0.3


appdirs==1.4.4
astroid==2.4.2
bcrypt==3.2.0
cffi==1.14.4

colorama==0.4.4
distlib==0.3.1
dnspython==1.16.0
filelock==3.0.12
Flask-Bcrypt==0.7.1
Flask-PyMongo==2.3.0
google-api-core==1.24.1
google-cloud==0.27.0
google-cloud-core==0.26.0
google-cloud-datastore==1.2.0
googleapis-common-protos==1.52.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycparser==2.20
pylint==2.6.0
pymongo==3.11.2
python-dotenv==0.15.0
six==1.15.0
virtualenv==20.2.2
wrapt==1.12.1